requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.19
django-filter==25.1
djangorestframework==3.15.2
redis==5.2.1
ruff==0.9.7
sqlparse==0.5.3
typing_extensions==4.12.2
uv==0.6.3
